serveur/storage/requirements.txt
flask
psycopg2-binary
langchain-core
langchain-text-splitters
langchain-qdrant
qdrant-client
langchain-ollama
litellm
nest_asyncio